third_party/WebKit/LayoutTests/fast/scroll-behavior/smooth-scroll/track-scroll-expected.txt

This test scrolls by clicking in the scrollbar track.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS scrollY became pageStep
PASS scrollY is pageStep
PASS successfullyParsed is true

TEST COMPLETE
